Fix the build of crypto_test on LP32 architectures
test->id is a uint64_t, not a long.

Sponsored by:	ConnectWise

+++ tests/zfs-tests/cmd/crypto_test.c
@@ -863,7 +863,8 @@ test_result(const crypto_test_t *test, int encrypt_rv, uint8_t *encrypt_buf,
 		return (pass);
 
 	/* print summary of test result */
-	printf("%s[%lu]: encrypt=%s decrypt=%s\n", test->fileloc, test->id,
+	printf("%s[%ju]: encrypt=%s decrypt=%s\n", test->fileloc,
+	    (uintmax_t)test->id,
 	    encrypt_pass ? "PASS" : "FAIL",
 	    decrypt_pass ? "PASS" : "FAIL");
